+++ app/component/Card.tsx
@@ -1,4 +1,6 @@
-function CardComponent(props) {
+import React from "react";
+
+function Card(props) {
   return (
     <div
       className={`flex rounded-lg border border-gray-200 bg-white shadow-md dark:border-gray-700 dark:bg-gray-800 flex-col md:w-1/2 py-2 px-3
@@ -9,4 +11,4 @@ function CardComponent(props) {
   );
 }
 
-export default CardComponent;
+export default Card;
