+++ app/helper/translation/bo.json
@@ -30,7 +30,5 @@
   "OCR_description": "པར་རིས་ནང་གི་བོད་ཡིག་ཡིག་འབེབས་བྱེད་ཐུབ།",
   "reset": "བསྐྱར་སྒྲིག",
   "submit": "ཐོངས།",
-  "jobs": "ས་མིག་གསལ་བསྒྲགས།",
-  "uploadImage": "འདིར་པར་རིས་འཇུག་རོགས།",
-  "acceptedImage": "ངོས་ལེན་ཡོད་པའི་པར་རྣམ།"
+  "jobs": "ས་མིག་གསལ་བསྒྲགས།"
 }
