+++ app/helper/translation/en.json
@@ -30,7 +30,5 @@
   "OCR_description": "Convert an image into text with AI",
   "reset": "Reset",
   "submit": "Submit",
-  "jobs": "Jobs",
-  "uploadImage": "Upload Image",
-  "acceptedImage": "Accepted Image"
+  "jobs": "Jobs"
 }
